Some memory optimizations, closes #1
Output from zdb from each zdb command is no longer loaded in memory but read
line by line to address memory usage. This change also make it so that the output
of this program is written as soon as every single file is processed.

Note though that we still load in memory the map counting the occurrences of the blocks,
which is probably the most memory intensive part, but to fix that I would need to
completely change the data structure and maybe the algorithm, so I will avoid
that unless it becomes an issue for somebody.